requirements.txt
tensorflow==2.15.0
opencv-python
numpy
pillow
